moved baseline correction from processing.py to standalone file. Modified adjust_highpass.py and processing.py to call correct_baseline()
