+++ gmprocess/waveform_processing/adjust_highpass.py
@@ -9,7 +9,7 @@ from gmprocess.waveform_processing.filtering import (
     lowpass_filter_trace,
     highpass_filter_trace,
 )
-
+from gmprocess.waveform_processing.baseline_correction import correct_baseline
 
 def adjust_highpass_corner(
     st,
@@ -93,21 +93,13 @@ def __disp_checks(tr, max_final_displacement=0.025, max_displacment_ratio=0.2):
     # Filter
     trdis = lowpass_filter_trace(trdis, **lp_args)
     trdis = highpass_filter_trace(trdis, **hp_args)
+    
+    # Apply baseline correction
+    trdis = correct_baseline(trdis)
 
     # Integrate to displacment
     trdis.integrate()
     trdis.integrate()
-
-    # Apply baseline correction
-    time_values = (
-        np.linspace(0, trdis.stats.npts - 1, trdis.stats.npts) * trdis.stats.delta
-    )
-    poly_cofs = list(curve_fit(_poly_func, time_values, trdis.data)[0])
-    poly_cofs += [0, 0]
-
-    # Construct a polynomial from the coefficients and subtract from displacement
-    polynomial = np.poly1d(poly_cofs)
-    trdis.data -= polynomial(time_values)
 
     # Checks
     ok = True
@@ -122,10 +114,3 @@ def __disp_checks(tr, max_final_displacement=0.025, max_displacment_ratio=0.2):
         ok = False
 
     return ok
-
-
-def _poly_func(x, a, b, c, d, e):
-    """
-    Model polynomial function for polynomial baseline correction.
-    """
-    return a * x ** 6 + b * x ** 5 + c * x ** 4 + d * x ** 3 + e * x ** 2
