+++ gmprocess/waveform_processing/processing.py
@@ -9,8 +9,6 @@ import numpy as np
 import logging
 
 from obspy.taup import TauPyModel
-from scipy.optimize import curve_fit
-from scipy.integrate import cumtrapz
 
 from gmprocess.core.stationtrace import PROCESS_LEVELS
 from gmprocess.core.streamcollection import StreamCollection
@@ -22,7 +20,7 @@ from gmprocess.waveform_processing.windows import (
 )
 from gmprocess.waveform_processing.phase import create_travel_time_dataframe
 from gmprocess.waveform_processing import corner_frequencies
-
+from gmprocess.waveform_processing.baseline_correction import correct_baseline
 # -----------------------------------------------------------------------------
 # Note: no QA on following imports because they need to be in namespace to be
 # discovered. They are not called directly so linters will think this is a
@@ -449,7 +447,7 @@ def detrend(st, detrending_method=None):
 
     for tr in st:
         if detrending_method == "baseline_sixth_order":
-            tr = _correct_baseline(tr)
+            tr = correct_baseline(tr)
         elif detrending_method == "pre":
             tr = _detrend_pre_event_mean(tr)
         else:
@@ -667,52 +665,3 @@ def _detrend_pre_event_mean(trace):
     noise_mean = np.mean(noise.data)
     trace.data = trace.data - noise_mean
     return trace
-
-
-def _correct_baseline(trace):
-    """
-    Performs a baseline correction following the method of Ancheta
-    et al. (2013). This removes low-frequency, non-physical trends
-    that remain in the time series following filtering.
-
-    Args:
-        trace (obspy.core.trace.Trace):
-            Trace of strong motion data.
-
-    Returns:
-        trace: Baseline-corrected trace.
-    """
-
-    # Integrate twice to get the displacement time series
-    disp_data = cumtrapz(
-        cumtrapz(trace.data, dx=trace.stats.delta, initial=0),
-        dx=trace.stats.delta,
-        initial=0,
-    )
-
-    # Fit a sixth order polynomial to displacement time series, requiring
-    # that the 1st and 0th order coefficients are zero
-    time_values = (
-        np.linspace(0, trace.stats.npts - 1, trace.stats.npts) * trace.stats.delta
-    )
-    poly_cofs = list(curve_fit(_poly_func, time_values, disp_data)[0])
-    poly_cofs += [0, 0]
-
-    # Construct a polynomial from the coefficients and compute
-    # the second derivative
-    polynomial = np.poly1d(poly_cofs)
-    polynomial_second_derivative = np.polyder(polynomial, 2)
-
-    # Subtract the second derivative of the polynomial from the
-    # acceleration trace
-    trace.data -= polynomial_second_derivative(time_values)
-    trace.setParameter("baseline", {"polynomial_coefs": poly_cofs})
-
-    return trace
-
-
-def _poly_func(x, a, b, c, d, e):
-    """
-    Model polynomial function for polynomial baseline correction.
-    """
-    return a * x ** 6 + b * x ** 5 + c * x ** 4 + d * x ** 3 + e * x ** 2
